malayalam-audio-transcriber/requirements.txt
openai-whisper
torch
sounddevice
soundfile
numpy
scipy
flask
werkzeug
flask-cors
